projects/langchain-indexing-experiment/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-indexing-experiment"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "faiss-cpu",
]
